+++ app/views/presse/show.html.haml
@@ -1,10 +1,10 @@
-- content_for(:head_title) { @article.titre }
+- content_for(:head_title) { @article.title }
 
 %main.fr-pt-2w.fr-pb-8w
   .fr-container
-    = render("shared/breadcrumbs", links: [["Accueil", root_path], ["Ils parlent de nous", presse_path]], current_page_label: @article.titre)
+    = render("shared/breadcrumbs", links: [["Accueil", root_path], ["Ils parlent de nous", presse_path]], current_page_label: @article.title)
 
-    %h1= @article.titre
+    %h1= @article.title
 
     .co-flex.co-flex--align-items-center.co-flex--gap-1rem
       = image_tag @article.logo, alt: "Logo #{@article.source}", style: "max-width: 150px; max-height: 80px;"
